requirements.txt
pytest>=7.0.0
requests>=2.28.0
python-dotenv>=0.20.0
pytest-html>=3.2.0
pytest-cov>=4.0.0
